prevents duplicate intersections
The current comparison duplicates the results. For example, if the sixth side of the polygon intersects with the ninth side, then the kinks will return both the intersections of 6 vs 9 as well as 9 vs 6.

This change iterates over ring2 starting not from the start but from the current ring1 index, because any prior intersection would have been already detected.